use longs to store lock timeout, and use BitConverter to serialize/deserialize
for improved efficiency.

+++ tests/ServiceStack.Redis.Tests/RedisClientTests.cs
@@ -361,7 +361,7 @@ namespace ServiceStack.Redis.Tests
             var key = "lockkey";
             var distributedLock = new DistributedLock();
             int lockTimeout = 2;
-            double lockVal;
+            long lockVal;
 
             Assert.AreNotEqual(lockVal = distributedLock.Lock(Redis,key,lockTimeout,lockTimeout),0);
